+++ changelog.md
@@ -8,6 +8,7 @@
 * Example Simple bus branch network added in `zepben.evolve.examples.simple_bus_branch`.
 * Helper factory methods for AcLineSegment, Junction (Bus), PowerTransformer, EnergySource, and EnergyConsumer added to `NetworkService` and available
 through `zepben.evolve.services.network.network_extensions`.
+* Added CIM Class LoadBreakSwitch
 
 ##### Enhancements
 * [tests] cim_creators.py created with similar hypothesis derived functionality to pb_creators, and can be used to quickly build fake data for testing.